edit exercise view and list deletions
